Added better main menu, now includes Options and Help, both WIP.
