+++ Blackjack/modules/options.py
@@ -1,0 +1,7 @@
+from . import menus
+from . import text_effect
+
+def main():
+    text_effect.slow_type("This is currently WIP. Returning to menu.")
+    text_effect.divide_lines
+    menus.main_menu()
