+++ Blackjack/todo.txt
@@ -1,5 +1,5 @@
 FEATURES / ITEMS TO WORK ON:
-- Add "help" to explain the game and special rules
+- Add "help" to explain the game and special rules (WIP)
 - Once Betting is available, understand more complex betting rules such as insurance and implement
 - Allow players to turn off betting
 - LONG TERM GOAL - Migrate to SQL database to hold player names and keep record of money long term. Need a way to add money if they run out,
